pyproject.toml
[project]
name = "meeting-schedule"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "logging>=0.4.9.6",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typing>=3.10.0.0",
]
